delete unused debug code
