+++ MMM/Sources/MMMApp.swift
@@ -20,17 +20,5 @@ struct MMMApp: App {
                     }
                 )
         }
-        .debugContainer()
-    }
-}
-
-private extension Scene {
-    @MainActor
-    func debugContainer() -> some Scene {
-        #if DEBUG
-            modelContainer(DebugSwiftData.container)
-        #else
-            self
-        #endif
     }
 }
