--- Package/Sources/AppDebug/Debug/DebugConfig.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-enum DebugConfig {
-    static var isTesting: Bool {
-        NSClassFromString("XCTestCase") != nil
-    }
-}
