+++ Package/Sources/AppDebug/DebugView/Development/DevelopmentView.swift
@@ -1,31 +1,11 @@
 import SwiftUI
 
 struct DevelopmentView: View {
-    @State private var isShowLogView = false
-
-    private var router = Router()
-
     var body: some View {
-        RouterView(router: router) {
-            Text("Development")
-                .toolbar {
-                    ToolbarItem(placement: .topBarTrailing) {
-                        Button("ログ履歴") {
-                            isShowLogView.toggle()
-                        }
-                    }
-                }
-                .sheet(
-                    isPresented: $isShowLogView,
-                    content: {
-                        NetworkingLogListView()
-                    }
-                )
-        }
+        Text("Development")
     }
 }
 
 #Preview {
     DevelopmentView()
-        .modelContainer(DebugSwiftData.previewContainer)
 }
