--- Package/Sources/AppDebug/Extension/View/Environment+Extension.swift
@@ -1,12 +1,0 @@
-import SwiftUI
-
-extension EnvironmentValues {
-    var router: Router {
-        get { self[RouterKey.self] }
-        set { self[RouterKey.self] = newValue }
-    }
-}
-
-private struct RouterKey: EnvironmentKey {
-    static let defaultValue = Router()
-}
